AnomalyDetectionInWebsiteTraffic/src/test/java/com/ml/AnomalyDetectionInWebsiteTraffic/service/AnomalyDetectionServiceTest.java
package com.ml.AnomalyDetectionInWebsiteTraffic.service;


import java.util.Arrays;
import java.util.List;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;  
import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertThrows;
import static org.mockito.Mockito.when;  


// Replace placeholders with your actual implementation details
@ExtendWith(MockitoExtension.class)
public class AnomalyDetectionServiceTest {

    /* WEKA filter object */
    @Mock
    private Object externalDependency;

    @Test
    public void testTrainModel() throws Exception {
        // Mock training exception
        when(externalDependency).thenThrow(new Exception("Training error"));
        AnomalyDetectionService service = new AnomalyDetectionService();
        assertThrows(Exception.class, () -> service.trainModel("data.csv"));
    }

    @Test
    public void testDetectAnomalies() throws Exception {
        // Mock successful anomaly detection
        List<Integer> anomalies = Arrays.asList(1, 3, 5);
        when(externalDependency).thenReturn(anomalies);

        AnomalyDetectionService service = new AnomalyDetectionService();
        List<Integer> detectedAnomalies = service.detectAnomalies("data.csv");

        assertEquals(anomalies, detectedAnomalies);
    }
}
